Fix _is_git_dirty method
Before this change this feature was broken and failing to run internal
if condition properly.

I think that the 'display' untracked files functionality is redundant here,
as the simplest thing we want to know is if repo has some non commited files
or not.

So, `git status --porcelain` is very simple replacement to check this.

Similar solution can be found in the other oh-my-fish theme:
https://github.com/oh-my-fish/theme-gitstatus